# Support nullable indices in boolean take kernel and some optimizations