--- a/arrow/src/compute/kernels/take.rs
+++ b/arrow/src/compute/kernels/take.rs
@@ -599,6 +599,40 @@ where
     Ok(PrimitiveArray::<T>::from(data))
 }
 
+fn take_bits<IndexType>(
+    values: &Buffer,
+    values_offset: usize,
+    indices: &PrimitiveArray<IndexType>,
+) -> Result<Buffer>
+where
+    IndexType: ArrowNumericType,
+    IndexType::Native: ToPrimitive,
+{
+    let len = indices.len();
+    let values_slice = values.as_slice();
+    let mut output_buffer = MutableBuffer::new_null(len);
+    let output_slice = output_buffer.as_slice_mut();
+
+    indices
+        .iter()
+        .enumerate()
+        .try_for_each::<_, Result<()>>(|(i, index)| {
+            if let Some(index) = index {
+                let index = ToPrimitive::to_usize(&index).ok_or_else(|| {
+                    ArrowError::ComputeError("Cast to usize failed".to_string())
+                })?;
+
+                if bit_util::get_bit(values_slice, values_offset + index) {
+                    bit_util::set_bit(output_slice, i);
+                }
+            }
+
+            Ok(())
+        })?;
+
+    Ok(output_buffer.into())
+}
+
 /// `take` implementation for boolean arrays
 fn take_boolean<IndexType>(
     values: &BooleanArray,
@@ -608,67 +642,12 @@ where
     IndexType: ArrowNumericType,
     IndexType::Native: ToPrimitive,
 {
-    let data_len = indices.len();
-
-    let mut val_buf = MutableBuffer::new_null(data_len);
-
-    let val_slice = val_buf.as_slice_mut();
-
-    let null_count = values.null_count();
-
-    let nulls = if null_count == 0 {
-        indices
-            .iter()
-            .enumerate()
-            .try_for_each::<_, Result<()>>(|(i, index)| {
-                if let Some(index) = index {
-                    let index = ToPrimitive::to_usize(&index).ok_or_else(|| {
-                        ArrowError::ComputeError("Cast to usize failed".to_string())
-                    })?;
-
-                    if values.value(index) {
-                        bit_util::set_bit(val_slice, i);
-                    }
-                }
-
-                Ok(())
-            })?;
-
-        indices.data_ref().null_buffer().cloned()
-    } else {
-        let mut null_buf = MutableBuffer::new_null(data_len);
-        let null_slice = null_buf.as_slice_mut();
-
-        indices
-            .iter()
-            .enumerate()
-            .try_for_each::<_, Result<()>>(|(i, index)| {
-                if let Some(index) = index {
-                    let index = ToPrimitive::to_usize(&index).ok_or_else(|| {
-                        ArrowError::ComputeError("Cast to usize failed".to_string())
-                    })?;
-
-                    if values.is_valid(index) {
-                        bit_util::set_bit(null_slice, i);
-                        if values.value(index) {
-                            bit_util::set_bit(val_slice, i);
-                        }
-                    }
-                }
-
-                Ok(())
-            })?;
-
-        match indices.data_ref().null_buffer() {
-            Some(buffer) => Some(buffer_bin_and(
-                buffer,
-                indices.offset(),
-                &null_buf.into(),
-                0,
-                indices.len(),
-            )),
-            None => Some(null_buf.into()),
+    let val_buf = take_bits(values.values(), values.offset(), indices)?;
+    let null_buf = match values.data().null_buffer() {
+        Some(buf) if values.null_count() > 0 => {
+            Some(take_bits(buf, values.offset(), indices)?)
         }
+        _ => None,
     };
 
     let data = unsafe {
@@ -676,9 +655,9 @@ where
             DataType::Boolean,
             indices.len(),
             None,
-            nulls,
+            null_buf,
             0,
-            vec![val_buf.into()],
+            vec![val_buf],
             vec![],
         )
     };